requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
six==1.17.0
reportlab